Move test_caffe2.c and harness to test/caffe2
