+++ tc/benchmarks/group_convolution.cc
@@ -26,7 +26,7 @@
 #include "tc/aten/aten_compiler.h"
 #include "tc/core/cuda/cuda_mapping_options.h"
 
-#include "../test/test_harness.h"
+#include "../test/caffe2/test_harness.h"
 #include "../test/test_harness_aten_cuda.h"
 #include "benchmark_fixture.h"
