+++ test/caffe2/CMakeLists.txt
@@ -1,0 +1,3 @@
+if (WITH_CUDA)
+  add_subdirectory(cuda)
+endif()
