+++ test/caffe2/cuda/CMakeLists.txt
@@ -1,0 +1,25 @@
+include_directories(.)
+include_directories(..)
+
+set(CMAKE_INSTALL_RPATH_USE_LINK_PATH TRUE)
+
+################################################################################
+# Don't ask ... TODO: cleanup gtest
+################################################################################
+include_directories(${PROJECT_SOURCE_DIR}/third-party/googletest/googletest/include)
+set(GTEST_LIBS gtest gtest_main)
+
+set(GOOGLE_LIBS ${GFLAGS_LIBRARIES} ${GLOG_LIBRARIES} ${GTEST_LIBS})
+
+################################################################################
+# Caffe2 + TC tests
+################################################################################
+set(TEST_FILES
+  test_caffe2
+)
+
+foreach(i ${TEST_FILES})
+  add_executable(${i} ${i}.cc)
+  add_test(${i} ${i})
+  target_link_libraries(${i} ${GOOGLE_LIBS} pthread tc_c2)
+endforeach()
